fix(vm): execution flow, refactor main

+++ src/vm/jump.c
@@ -6,7 +6,7 @@
 /*   By: smamalig <smamalig@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/02 22:12:21 by smamalig          #+#    #+#             */
-/*   Updated: 2025/10/02 22:15:55 by smamalig         ###   ########.fr       */
+/*   Updated: 2025/10/05 20:32:07 by smamalig         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ void	vm_jz(t_vm *vm, t_program *program)
 	if (ft_vector_at(&vm->exit_codes, -1).value.i32 != 0)
 		return ;
 	ft_memcpy(&jmp_pos, program->data + program->pc + 1, sizeof(int32_t));
-	program->pc = jmp_pos - 1;
+	program->pc = (size_t)jmp_pos - 1;
 }
 
 void	vm_jnz(t_vm *vm, t_program *program)
@@ -31,5 +31,5 @@ void	vm_jnz(t_vm *vm, t_program *program)
 	if (ft_vector_at(&vm->exit_codes, -1).value.i32 == 0)
 		return ;
 	ft_memcpy(&jmp_pos, program->data + program->pc + 1, sizeof(int32_t));
-	program->pc = jmp_pos - 1;
+	program->pc = (size_t)jmp_pos - 1;
 }
